api: Add pull mode support for HttpNfcLease
vSphere 6.7 and later have support for upgrading the HttpNfcLease to use
pull mode if supported by the ESX host. In order to determine if the
lease is eligible to be upgraded, the capabilities of the HttpNfcLease
must be checked. If the lease is eligible for upgrading, a new
pullFromURLs task upgrades the lease and the ESXI host will fetch the
necessary file(s) directly from the URLs instead of relying on the
client to upload them.

In order to support this functionality, the nfc Lease object has new
GetCapabilities, PullFromUrls_Task, Upgrade, and WaitForPull functions
added. When the Upgrade function is called, it will use GetCapabilities
to determine if the lease is eligible to be upgraded or not. If it is
not, an error is returned. If it is, all FileItems found in the lease
are converted to HttpNfcLeaseSourceFiles that are passed to the
PullFromUrls_Task function along with the SSL thumbprint of the host to
download the files from to upgrade the lease to use pull mode. The new
WaitForPull function monitors the transferProgress until all
HttpNfcLeaseSourceFiles have been pulled by the ESXI host.

The Importer now has a PullMode and Thumbprint as part of its attributes.
When PullMode is set to true, the Importer will call the new Upgrade
function on the nfc Lease object. It then uses the new WaitForPull
function on the nfc Lease object to wait until the pull of all files has
completed before marking the nfc Lease as complete.