govc: Add support for pull mode on imports
The import.ovf and import.ova commands now both have new pull-mode and
thumbprint arguments. When pull-mode is specified, the SSL thumbprint of
the host to pull the file(s) from must be specified using the thumbprint
argument. When the import is started, the HttpNfcLease will be upgraded
to pull mode if its capabilities indicate it is supported. This tells
the ESXI host to pull the files directly from the remote URL instead of
relying on the client to upload the files to the ESXI host.